chore: move more schemas
