+++ ui/src/schemas/members.ts
@@ -1,0 +1,11 @@
+// Libraries
+import {schema} from 'normalizr'
+
+// Types
+import {ResourceType} from 'src/types'
+
+/* Members */
+
+// Defines the schema for the "members" resource
+export const memberSchema = new schema.Entity(ResourceType.Members)
+export const arrayOfMembers = [memberSchema]
